backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
Faker==19.13.0
python-multipart==0.0.6
typing-extensions==4.8.0
starlette==0.27.0 